scaffolder-backend: split builtin actions into multiple modules
