+++ plugins/scaffolder-backend/src/scaffolder/actions/builtin/fetch/index.ts
@@ -14,6 +14,5 @@
  * limitations under the License.
  */
 
-export * from './builtin';
-export { TemplateActionRegistry } from './TemplateActionRegistry';
-export type { ActionContext, TemplateAction } from './types';
+export { createFetchPlainAction } from './plain';
+export { createFetchCookiecutterAction } from './cookiecutter';
